pyproject.toml
[tool.poetry]
name = "ptb-ext"
version = "0.1.1"
description = "Useful helper functions & extensions for the python-telegram-bot framework"
authors = ["lurian <lurian-code@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {extras = ["all"], version = "^20.1"}
pydantic = "^2.5.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
